% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DSS_TS.R
\name{getDSS}
\alias{getDSS}
\title{getDSS}
\usage{
getDSS(dssFile, zone, component)
}
\arguments{
\item{dssFile}{a DSS file that you've opened with DSSRip}

\item{zone}{IWFM zone budget zone number}

\item{component}{the name of the flow component to import; part F of the DSS file}
}
\description{
This function gets returns a single time series from a DSS file generated by the IWFM zone budget utility
}
\examples{
getDSS()
}
\keyword{dss}
\keyword{iwfm}
